Option: use only import / export steps explicitly names by the current profile.
If this option is set, the setup tool ignores any steps which might be
globally registered:  it uses only those named in the profile's
'import_steps.xml' / 'export_steps.xml' files.

This is particularly useful for configuring sites with baseline profiles,
where arbitrary add-on steps are not only useless, but potentially damaging.